requirements.txt
Django
Pillow
python-decouple
django-allauth
django-crispy-forms==1.14.0
psycopg2-binary
django-extensions
#uWSGI
django-localflavor